+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 3.55.0
+  - feat: Add project migration publisher with module status tracking
+
+# 3.54.2
+  - fix: Adding missing dynamo db ambient variables
+
+# 3.54.1
+  - feat: Implement internal staff email bypass for SSO enforcement
+
+# 3.54.0
+  - feat: Add get token and invalidate token endpoint, adding weni-commons version 1.3.2a12, and token creating and invalidation use cases.
+
 # 3.53.0
   - feat: Add AMQ Project Publisher